+++ crates/swc_ecma_compat_common/src/regexp.rs
